+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - feature: install cursor cli and gui
 - feature: install frame0
 - feature: install hermes agent
+- feature: install insomnia
 
 ## [1.4.0] - 2026-07-24